requirements.txt
PyQt6
qrcode[pil]
Pillow

